pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "django-advanced"
description = "Django Advanced"
version = "0.1.0"
readme = "README.md"
authors = ["localhost"]


[tool.poetry.dependencies]
python = "3.9.7"

django = "^3.2"
django-split-settings = "^1.1"
django-axes = "^5.27"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.5"
django-stubs-ext = "^0.3"

psycopg2-binary = "^2.9"
gunicorn = "^20.0"
python-decouple = "^3.5"
bcrypt = "^3.2"
structlog = "^21.3"
django-autoslug = "^1.9.8"
Pillow = "^8.4.0"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
django-querycount = "^0.7"
django-migration-linter = "^3.0"
django-extra-checks = "^0.11"
django-coverage-plugin = "^2.0"
nplusone = "^1.0"

wemake-python-styleguide = "^0.15"
flake8-pytest-style = "^1.5"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
nitpick = "^0.29"

pytest = "^6.2"
pytest-django = "^4.4"
pytest-cov = "^3.0"
pytest-randomly = "^3.10"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.2"
pytest-timeout = "^2.0"
django-test-migrations = "^1.1"
hypothesis = "^6.27"

mypy = "^0.910"
django-stubs = "^1.9"

sphinx = "^4.3"
sphinx-autodoc-typehints = "^1.12"
tomlkit = "^0.7"
doc8 = "^0.10"

yamllint = "^1.26"
safety = "^1.10"
dotenv-linter = "^0.2"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.3"
ipython = "^7.29"
